requirements.txt
fastapi
uvicorn[standard]
python-dotenv
transformers
torch
pydantic 